+++ src/cf-atomic-component/src/utilities/transition/BaseTransition.js
@@ -1,6 +1,5 @@
 // Required modules.
 const Events = require( '../../mixins/Events.js' );
-const fnBind = require( '../function-bind' ).bind;
 
 /* eslint-disable max-lines-per-function, max-statements */
 /**
@@ -32,8 +31,8 @@ function BaseTransition( element, classes ) {
    * @returns {BaseTransition} An instance.
    */
   function init() {
-    _transitionCompleteBinded = fnBind( _transitionComplete, this );
-    _addEventListenerBinded = fnBind( _addEventListener, this );
+    _transitionCompleteBinded = _transitionComplete.bind( this );
+    _addEventListenerBinded = _addEventListener.bind( this );
     setElement( element );
 
     return this;