+++ CHANGELOG.md
@@ -17,6 +17,7 @@ We follow the [Semantic Versioning 2.0.0](http://semver.org/) format.
 - **cf-tables:** [PATCH] Updates test simulate event syntax to es6 modules.
 
 ### Removed
+- **cf-atomic-component:** [PATCH] Removes deprecated function bind polyfill.
 - **cf-expandables:** [MINOR] Remove classlist-polyfill for IE9.
 - **cf-atomic-component:** [MINOR] Remove dom-class-list module for IE9.